requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
urllib3==2.0.7
dnspython==2.4.2
colorama==0.4.6
tqdm==4.66.1
argparse==1.4.0
json-logging==1.5.0
concurrent.futures==3.1.1
webdriver-manager==4.0.1
lxml==4.9.3
